Move Release class to Models Folder
